requirements.txt
pandas==2.2.3
numpy==2.2.4
scipy==1.15.2
openai==1.77.0
httpx==0.28.1
pydantic==2.11.3
python-dotenv==1.1.0
tabulate==0.9.0
tqdm==4.67.1
nltk==3.9.1
requests==2.32.3
matplotlib-inline==0.1.7
krippendorff==0.8.1
